requirements.txt
opencv-python
argparse
urllib3
numpy
mss
Pillow
bluepy
pygatt
apscheduler
pexpect
Flask